Ongoing with CalculateQuotation method

+++ QuoteCalculator.Domain/Models/FinanceModel.cs
@@ -3,7 +3,7 @@
     public class FinanceModel
     {
         public Guid Id { get; set; } = new Guid();
-        public int FinanceAmount { get; set; }
+        public decimal FinanceAmount { get; set; }
         public required string RepaymentSchedule { get; set; }
         public required QuotationModel Quotation { get; set; }
         public required ProductModel Product { get; set; }
